getpost and get postS rewrited

+++ internal/models/user.go
@@ -1,8 +1,8 @@
 package models
 
 type User struct {
-	User_id       int    `db:"user_id"`
-	User_lvl      int    `db:"user_lvl"`
+	User_id       int64  `db:"user_id"`
+	User_lvl      int64  `db:"user_lvl"`
 	User_email    string `db:"user_email"`
 	User_nickname string `db:"user_nickname"`
 }
